% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_chr_scalar}
\alias{check_chr_scalar}
\title{Vérifer ce que l'utilisateurs donne}
\usage{
check_chr_scalar(
  x,
  complement = NULL,
  call = rlang::caller_env(),
  arg = rlang::caller_arg(x),
  ...
)
}
\arguments{
\item{x}{l'objet à vérifer}

\item{complement}{Un complément d'information pour aguiller l'utilisaeur
(optionnel)}

\item{call}{Pour que l emessage d'erreur mentionne la fonction que
l'utilisateur a appelé}

\item{arg}{Pour choper le nom de l'argument dans lequel la fonction a été
utilisé.}

\item{classe}{le \if{html}{\out{<chr>}} qui identifie la classe}

\item{message}{Le message d'erreur à afficher}
}
\value{
Rien du tout, mais un message d'erreur si ça foire
}
\description{
Vérifer ce que l'utilisateurs donne
}
\examples{

# Tests
bidule <- function(a) {
  a |> check_inherits("data.frame", message2 = c("i" = "Avez vous pensé à laver ? "))
}

bidule("pates")




}
