% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taux_moyen_theorique.R
\name{contribution_moyenne_theorique}
\alias{contribution_moyenne_theorique}
\title{Calculer la contribution moyenne théorique}
\usage{
contribution_moyenne_theorique(resultat_acm)
}
\arguments{
\item{resultat_acm}{Le résultat d'une ACM}
}
\value{
un nombre entre 100 et 0
}
\description{
Certains manuels recommandent de calculer la contribution moyenne théorique pour s'en servir comme seuil afin de déterminer quelles sont les variables à prendre en compte dans l'analyse. Cette fonction permet de calculer cette contribution moyenne théorique : elle fait 100 \% divisé par le nombre de modalité active.
}
\examples{
resultat_acm <- createurs_de_contenus_pol |>
  FactoMineR::MCA(quali.sup = c("titre", "personne", "orientation"))

contribution_moyenne_theorique(resultat_acm)


}
