% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_acm.R
\name{coord_acm}
\alias{coord_acm}
\title{Title}
\usage{
coord_acm(...)
}
\arguments{
\item{...}{une liste optionnelle d'argument à passer à [ggplot2::coord_fixed()]. Cet argument est optionel}
}
\value{
un objet ggplot qui sert de canva pour les ACM
}
\description{
Title
}
\examples{
# une série d'exemple
}
