% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraire_classes.R
\name{extraire_classes}
\alias{extraire_classes}
\title{Extraire les résultats d'une classification ascendante hiérarchique}
\usage{
extraire_classes(resultat_classification, ...)
}
\arguments{
\item{resultat_classification}{Un objet qui contient le résultat d'une
classification réalisée par \code{\link[FactoMineR:HCPC]{FactoMineR::HCPC()}}.}

\item{...}{Colonnes placées en début de tableau. Argument inutile.}
}
\value{
Le tableau de donnée initial utilisé pour l'ACM auquel a été rajouté
une colonne \code{classe} qui renseigne la classe des individus. Les colonnes
de résultat de l'ACM telle que décrite dans \code{\link[=extraire_individus]{extraire_individus()}} sont
également incluses.
}
\description{
Lorsque l'on réalise une classification ascendante hiérarchique, et que
l'on souhaite utiliser les résultats, on a souvent envie d'avoir ces résultats
acollés à notre tableau d'origine.
}
\details{
Cette fonction permet d'extraire d'un résultat d'une classification réalisé
avec \code{\link[FactoMineR:HCPC]{FactoMineR::HCPC()}} le tableau de données d'origines enrichi d'une
colonne qui précise à quelle classe appartient chaque individus.

Le jeu de données comporte également les colonnes des résultats de l'ACM.
Pour plus de détails sur ces colonnes, voir \code{\link[=extraire_individus]{extraire_individus()}}.

Le principal intérêt de cette fonction est de permettre de générer un tableau
prêt à être utilisé par \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} pour en faire un graphique.
}
\examples{

# 1. On réalise une ACM avec ACM() ou FactoMineR::MCA()
resultat_acm <- createurs_de_contenus_pol |> ACM()

# 2. On fait la classification sur le résultat de l'ACM
resultat_classification <- resultat_acm |> FactoMineR::HCPC(nb.clust = 2)

# 3. On peut extraire du résultat de la classification les classes.
resultat_classification |> extraire_classes()
}
