% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_individus.R
\name{extraire_individus}
\alias{extraire_individus}
\title{Obtenir les résultats pour chaque individus}
\usage{
extraire_individus(resultat_acm)
}
\arguments{
\item{resultat_acm}{resultat acm}
}
\value{
Le tableau de données surlequel vous avez réalisé l'ACM auquel a été
 ajouté pour chaque individus : leur coordonnées sur chaque axe (stockés dans des
 colonnes `dim1_coord`, `dim2_coord`, `dim3_coord`, etc), leur contribution à
 la construction des axes (colonnes `dim1_contrib`, `dim2_contrib`,...) et
 enfin leur cosinus carré (colonnes `dim1_cos2`,...)
}
\description{
Cette fonction permet quadd on lui donne le résultat d'une ACM de sortir un
tableau qui reprend le tableau de donnée initial en ajoutant pour chaque
individus leur coordonnées.
}
\examples{
resultat_acm <- createurs_de_contenus_pol |> ACM()

resultat_individus <- resultat_acm |> extraire_individus()

resultat_individus
}
