% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variance.R
\name{extraire_variance}
\alias{extraire_variance}
\title{Décrire la variance des capturer par chaque axe}
\usage{
extraire_variance(resultat_acm)
}
\arguments{
\item{resultat_acm}{un résultat d'ACM qui sort de factominer::MCA()}
}
\value{
un tableau (data.frame) qui documente la variance des 5 premiers axes.
}
\description{
Décrire la variance des capturer par chaque axe
}
\examples{
# une liste d'exemple
}
