% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frequence.R
\name{get_frequence}
\alias{get_frequence}
\title{Obtenir la fréquence des modalités}
\usage{
get_frequence(data_acm)
}
\arguments{
\item{data_acm}{le data.frame à qui a été donné à manger à Factominer}
}
\value{
un tableau qui renvoie pour chaque modalité le nombre d'individu concerné (n) ainsi que leur pourcentage
}
\description{
Obtenir la fréquence des modalités
}
\examples{
# une liste d'exemple
}
