% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_res.R
\name{get_res}
\alias{get_res}
\title{Title}
\usage{
get_res(resultat_acm_var, data_acm, active = TRUE)
}
\arguments{
\item{resultat_acm_var}{Résultat de l'ACM avec Factominer}

\item{data_acm}{Tableau original}

\item{active}{TRUE ou FALSE. TRUE pour extraire les variables actives. FALSE pour extraire les variables supplémentaires}
}
\value{
pour chaque modalité les coordonnées, les contributions, les vtest et les cosinus carrés pour les 5 premiers axes.
}
\description{
Title
}
\examples{
# une liste d'exemple
}
