% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggacm.R
\name{ggacm}
\alias{ggacm}
\title{Visualiser les résultats d'une ACM}
\usage{
ggacm(data, mapping = aes(dim1_coord, dim2_coord, label = modalites), ...)
}
\arguments{
\item{data}{Le résultat d'une ACM après qu'il soit passé par la fonction [get_individus()] ou [get_variables()].}

\item{mapping}{Par défaut prend les coordonnée de la dimension 1 en abcisse (`x = dim1coord`) et de la dimension 2 en ordonné (`y = dim2coord`), les textes sont les modalités (`label = modalites`). Il est possible d'utiliser d'autres aesthetics. Pour plus d'information voir la vignette et la documentation de ggplot. Cet argument doit toujours être écrit avec la fonction [aes()] comme : `aes(x = ..., y = ..., label = ..., etc.)`}

\item{...}{Arguments optionnels passés à [coord_acm()] qui seront passé à [ggplot2::coord_fixed()].}
}
\value{
un graphique ggplot
}
\description{
Visualiser les résultats d'une ACM
}
\examples{
# un exemple pas piqué des annetons
}
