% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mise_en_place.R
\name{mise_en_place_ACM}
\alias{mise_en_place_ACM}
\title{Mettre en place une session R adapté}
\usage{
mise_en_place_ACM(...)
}
\arguments{
\item{...}{Il est possible de faire passer d'autres arguments à \code{\link[base:options]{base::options()}}.}
}
\value{
Une liste d'option
}
\description{
R affiche par défaut beaucoup trop de chiffres après la virgule
(qui est d'ailleurs un point en général) et tend à utiliser l'écriture
scientifique... ce qui n'arrange pas vraiment les chercheurs en sciences
sociales... Cette fonction permet de modifier ces options d'un seul coup.
Un nombre qui s'affichait comme 0.456789 s'affichera alors comme 0,457.
}
\details{
Concrètement cette fonction modifie les options de la session R et les options
des objets du package gtsummary en ayant recours à \code{\link[gtsummary:theme_gtsummary]{gtsummary::theme_gtsummary_language()}}.
}
\examples{

# ne fait pas parti  de l'exemple
old <- options()

# Exemple :

# Avant :
c(0.123456, 10e20)

mise_en_place_ACM()

# Après :
c(0.123456, 10e20)

# Fin de l'exemple
options(old)

}
