% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nommer_axes.R
\name{nommer_axes}
\alias{nommer_axes}
\title{Nommer les axes d'une représentation graphique d'ACM}
\usage{
nommer_axes(
  resultat_acm,
  en_haut = "",
  en_bas = "",
  a_gauche = "",
  a_droite = ""
)
}
\arguments{
\item{resultat_acm}{le résultat d'une ACM sortie de MCA()}

\item{en_haut}{à quoi correspond le haut du graphique ? entre guillemet.}

\item{en_bas}{idem mais en bas}

\item{a_gauche}{idem mais à gauche}

\item{a_droite}{idem mais a droite}
}
\value{
une couche ggplot à ajouter. Sous le capot cette fonction fait génère des annotations pour graphique ggplot avec \code{\link[ggplot2:labs]{ggplot2::labs()}}
}
\description{
Lorsque l'on représente une ACM, il est courant de vouloir donner un nom aux axes, ou au moins les pôles de chacun de ces axes. Cette fonction permet de générer les labels pour chaque axe à utiliser lorsqu'on génère des graphiques axec ggplot
}
\examples{
#
}
