% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuage_individus.R
\name{nuage_individus}
\alias{nuage_individus}
\title{Visualiser rapidement le nuage d'individus}
\usage{
nuage_individus(resultat_acm, couleur)
}
\arguments{
\item{resultat_acm}{Le résultat d'une acm}

\item{couleur}{on peut optionnellement coloré le nuage d'individus en fonction d'une variable.}
}
\value{
un objet ggplot
}
\description{
On veut souvent visualiser rapidement un nuage d'individus sans trop y passer du temps. Cette fonction est un raccourci à cela, elle permet de faire
}
\examples{
#
}
