% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuage_individus_facette.R
\name{nuage_individus_facette}
\alias{nuage_individus_facette}
\title{Visualiser les positions des individus en fonction des modalités}
\usage{
nuage_individus_facette(
  resultat_individus,
  variables_a_representer,
  forme = 17,
  couleur = "#e41a1c",
  longueur_max_labels = 20,
  ...
)
}
\arguments{
\item{resultat_individus}{Un tableau de donnée avec les coordonnées des individus
qui a été produit par extraire_individus().
La fonction prend également en charge un résultat d'ACM directement si besoin.}

\item{variables_a_representer}{La liste des variables à représenter :
on si on souhaite sélectionner le \code{genre}, la \code{classe} et le \code{niveau_de_revenu}
on peut par exemple écire \code{c(genre, classe, niveau_de_revenu)}}

\item{forme}{Quelles formes pour les individus qui partagent une même modalité ?
Par défaut un triangle. Doit-être un nombre, voir la documentation de ggplot
sur les "shapes" pour les possibilités.}

\item{couleur}{Quelle couleur doit-être utilisée ?
doit-être une chaiîne de caractère qui encode la couleur comme "#e41ac" ou "yellow"
par exemple.}

\item{longueur_max_labels}{Longeur maximum des labels de chaque facette.
Par défaut chaque label ira à la ligne au bout de 20 caractères.}

\item{...}{Arguments supplémentaires pour les utilisateurs avancés
qui seront passés à ggplot2::facet_wrap()}
}
\value{
Un objet ggplot2.
}
\description{
Cette fonction permet de faire apparaître plusieurs fois le nuage d'individu en colorant à chaque fois les individus qui partagent une même modalité.
}
\examples{

res.ACM <- createurs_de_contenus_pol |> ACM()

res.ind <- res.ACM |> extraire_individus()

res.ind |>
nuage_individus_facette(variables_a_representer = c(CNC, partenariat_entreprise))
}
