% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supprimer_var_.R
\name{supprimer_var_}
\alias{supprimer_var_}
\title{Supprimer le format variable_modalite}
\usage{
supprimer_var_(df)
}
\arguments{
\item{df}{un tableau à nettoyer}
}
\value{
Le data.frame nettoyé
}
\description{
FactoMineR renvoie parfois pour certaines colonnes un résultat avec des modalités qui ont la forme de "variable_modalite". Cette fonction permet de nettoyer les tableaux qui se présentent sous cette forme.
}
\examples{
#
}
