% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caracteriser_classe.R, R/tableau_classe.R
\name{tableau_classe}
\alias{tableau_classe}
\title{Caractériser une classification}
\usage{
tableau_classe(classification, n_classe, numero_classe)

tableau_classe(classification, n_classe, numero_classe)
}
\arguments{
\item{classification}{une classification}

\item{n_classe}{Le nombre de personnes présentes dans la classe}

\item{numero_classe}{Le numéro de la classe}

\item{x}{Un objet qui contient le résultat d'une classification réalisée avec FactoMineR::HCPC().}

\item{data_acm}{Le tableau de donnée qui sur lequel l'ACM a été réalisée.}
}
\value{
Un tableau prêt à l'exportation décrivant une classe.

Un
}
\description{
Cette fonction permet de présenter un tableau caractérisant une classification réalisé avec la fonction FactoMineR::HCPC()
}
\examples{
# un exemple

# un exemple
}
