% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_acm.R
\name{theme_acm}
\alias{theme_acm}
\title{Utiliser un thème adapté à l'analyse de correspondance multiple}
\usage{
theme_acm(...)
}
\arguments{
\item{...}{Si vous voulez customiser le thème plus en profondeur, il est possible de rajouter des éléments qui seront passé à \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
Un thème ggplot
}
\description{
Cette fonction permet de générer un thème de graphique adapté aux ACM et aux publications : Fond blanc, police Times New Roman, légende en bas du graphique et transforme les axes en flèches.
}
\examples{
#
}
