% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau_axe.R
\name{caracteriser_axe}
\alias{caracteriser_axe}
\title{Caractériser ses axe
La fonction suivante permet de sortir un tableau résumant pour un axe les modalités les plus contribuantes à sa construction.}
\usage{
caracteriser_axe(resultat_acm, axe, contribution_minimum)
}
\arguments{
\item{resultat_acm}{Un résultat d'ACM de \code{\link[FactoMineR:MCA]{FactoMineR::MCA()}} ou \code{\link[=ACM]{ACM()}}.}

\item{axe}{Numéro de l'axe à décrire. (1, 2, 3, ...)}

\item{seuil}{Seuil de la contribution à l'axe en dessous duquel les modalités ne seront pas prises en compte. Vaut 0 par défaut, mais il peut être utile de l'augmenter si l'on a beaucoup de modalité et que l'on ne souhaite pas qu'elles apparaissent toutes dans le tableau.}
}
\value{
Un tableau de donnée prêt à être publié. Il s'agit d'une objet venant du package gt.
}
\description{
Caractériser ses axe
La fonction suivante permet de sortir un tableau résumant pour un axe les modalités les plus contribuantes à sa construction.
}
\examples{
#
}
