% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_individus.R
\name{extraire_individus}
\alias{extraire_individus}
\title{Obtenir les résultats pour chaque individus}
\usage{
extraire_individus(resultat_acm, data_acm, nom_variable_individu)
}
\arguments{
\item{resultat_acm}{resultat acm}

\item{data_acm}{donnée acm}

\item{nom_variable_individu}{une chaîne de caractère (qui s'écrit donc entre guillemets anglais "") qui donne le nom de la variable du tableau qui identifie les individus.}
}
\value{
un tableau data.frame au format tidy, prêt à être utilisé pour ggplot
}
\description{
Obtenir les résultats pour chaque individus
}
\examples{
# un exemple pas piqué des annetons
}
