% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuage_individus.R
\name{nuage_individus}
\alias{nuage_individus}
\title{Visualiser rapidement le nuage d'individus}
\usage{
nuage_individus(resultat_individus, couleur)
}
\arguments{
\item{couleur}{on peut optionnellement coloré le nuage d'individus en fonction d'une variable.}

\item{resultat_acm}{Le résultat d'une acm}
}
\value{
un objet ggplot
}
\description{
On veut souvent visualiser rapidement un nuage d'individus sans trop y passer du temps. Cette fonction est un raccourci à cela, elle permet de faire
}
\examples{

# on commence par créer un résultat d'ACM :
resultat <- createurs_de_contenus_pol |>
  ACM(variables_supplementaires = personne)

# puis on peut sortir le graphique
resultat |>
  nuage_individus() +
  theme_acm()

}
