% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caracteriser_classe.R
\name{tableau_classe}
\alias{tableau_classe}
\title{Caractériser une classification}
\usage{
tableau_classe(x, n_classe, data_acm)
}
\arguments{
\item{x}{Un objet qui contient le résultat d'une classification réalisée avec FactoMineR::HCPC().}

\item{n_classe}{Le numéro de la classe que l'on souhaite caractériser}
}
\value{
Un tableau prêt à l'exportation décrivant une classe.
}
\description{
Cette fonction permet de présenter un tableau caractérisant une classification réalisé avec la fonction FactoMineR::HCPC()
}
\examples{
# un exemple

}
